refactor: move Navbar logic into Navbar folder
