+++ wealth-sync/src/components/Navbar/SignUpButton.tsx
